Move offline BatCache to client package
